Update linear_algebra_calculator_3.1.py
added more to the error handling when getting the number of rows for calculating the determinant of a matrix.
added the calculate determinant function and used the .get function to get the rows so we have a matrix with all the entries for the determinant.
renamed the frame to get_det_matrix_frame